refactor(server): split iot worker runtime
